Refactor datapoints into separate file
